requirements.txt
tensorflow==2.16.1
matplotlib==3.8.4
ipython==8.22.2
numpy==1.22.3
streamlit==1.33.0
pillow==10.3.0